app/requirements.txt
oslo.messaging==9.7.1
python-cinderclient==4.2.0
python-swiftclient===3.8.0
django==2.1.7
django-bootstrap3==11.0.0
django-cors-headers==2.5.2
jsonpickle==1.1
six==1.12.0
requests==2.21.0
djangorestframework==3.9.1
apscheduler==3.6.0
jinja2==2.10
drf_yasg==1.15.0
django-rest-swagger==2.2.0
flex
celery==4.2.0
django-celery==3.2.2
mysqlclient==1.4.4
swagger_spec_validator
openstacksdk==0.27.0
six==1.12.0
keystoneauth1==3.14.0
python-keystoneclient==3.19.0
werkzeug==0.15.4
python-memcached==1.59
django-requestlogging-redux==1.2.1
pycryptodome==3.8.1
PyJWT==1.7.1
anyjson==0.3.3
python-barbicanclient==4.9.0
confluent-kafka==1.4.1
